Feature:removed the delete button from the student. Apply the delete student on the main menu for teacher only.
